pyproject.toml
[project]
name = "imfia-experiments"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib (>=3.10.1,<4.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
mypy = "^1.15.0"
python-dotenv = "^0.18.0"
isort = "^6.0.1"
ruff = "^0.11.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
